# Hotfix/memory allocation