--- a/README.md
+++ b/README.md
@@ -49,23 +49,23 @@ import (
 
 func main() {
         // client that tries to connect to statsd service, fallback to debug log backend if fails to connect
-        statsdClient, _ := stats.NewClient("statsd://statsd-host:8125", "my.app.prefix")
+        statsdClient, _ := stats.NewClient("statsd://statsd-host:8125/?prefix=my.app.prefix")
         defer statsdClient.Close()
 
         // debug log backend for stats
-        logClient, _ := stats.NewClient("log://", "")
+        logClient, _ := stats.NewClient("log://")
         defer logClient.Close()
 
         // memory backend to track operations in unit tests
-        memoryClient, _ := stats.NewClient("memory://", "")
+        memoryClient, _ := stats.NewClient("memory://")
         defer memoryClient.Close()
 
         // noop backend to ignore all stats
-        noopClient, _ := stats.NewClient("noop://", "")
+        noopClient, _ := stats.NewClient("noop://")
         defer noopClient.Close()
 
         // get settings from env to determine backend and prefix
-        statsClient, _ := stats.NewClient(os.Getenv("STATS_DSN"), os.Getenv("STATS_PREFIX"))
+        statsClient, _ := stats.NewClient(os.Getenv("STATS_DSN"))
         defer statsClient.Close()
 }
 ```
@@ -208,7 +208,7 @@ import (
 )
 
 func TestDoSomeJob(t *testing.T) {
-        statsClient, _ := stats.NewClient("memory://", "") 
+        statsClient, _ := stats.NewClient("memory://") 
 
         err := DoSomeJob(statsClient)
         assert.Nil(t, err)
@@ -271,7 +271,7 @@ func main() {
         if err != nil {
                 sectionsTestsMap = map[bucket.PathSection]bucket.SectionTestDefinition{}
         }
-        statsClient, _ := stats.NewClient(os.Getenv("STATS_DSN"), os.Getenv("STATS_PREFIX"))
+        statsClient, _ := stats.NewClient(os.Getenv("STATS_DSN"))
         statsClient.SetHTTPMetricCallback(bucket.NewHasIDAtSecondLevelCallback(&bucket.SecondLevelIDConfig{
                 HasIDAtSecondLevel:    sectionsTestsMap,
                 AutoDiscoverThreshold: 25,