--- a/bucket/metric_storage_test.go
+++ b/bucket/metric_storage_test.go
@@ -12,11 +12,11 @@ func TestMetricStorage_LooksLikeID(t *testing.T) {
 	firstSection := time.Now().Format(time.RFC3339Nano)
 
 	for i := uint(0); i < storage.threshold-1; i++ {
-		assert.False(t, storage.LooksLikeID(firstSection, time.Now().Format(time.RFC3339Nano)))
+		assert.False(t, storage.LooksLikeID(firstSection, string(i)))
 	}
 
-	assert.True(t, storage.LooksLikeID(firstSection, time.Now().Format(time.RFC3339Nano)))
-	assert.True(t, storage.LooksLikeID(firstSection, time.Now().Format(time.RFC3339Nano)))
+	assert.True(t, storage.LooksLikeID(firstSection, string(storage.threshold+1)))
+	assert.True(t, storage.LooksLikeID(firstSection, string(storage.threshold+1)))
 
 	assert.Equal(t, storage.threshold, uint(len(storage.metrics[firstSection])))
 }