--- a/client_test.go
+++ b/client_test.go
@@ -7,26 +7,26 @@ import (
 )
 
 func TestNewClient(t *testing.T) {
-	client, err := NewClient("statsd://", "")
+	client, err := NewClient("statsd://")
 	assert.NoError(t, err)
 	assert.IsType(t, &StatsdClient{}, client)
 
 	statsdClient, _ := client.(*StatsdClient)
 	assert.True(t, statsdClient.muted)
 
-	client, err = NewClient("log://", "")
+	client, err = NewClient("log://")
 	assert.NoError(t, err)
 	assert.IsType(t, &LogClient{}, client)
 
-	client, err = NewClient("memory://", "")
+	client, err = NewClient("memory://")
 	assert.NoError(t, err)
 	assert.IsType(t, &MemoryClient{}, client)
 
-	client, err = NewClient("noop://", "")
+	client, err = NewClient("noop://")
 	assert.NoError(t, err)
 	assert.IsType(t, &NoopClient{}, client)
 
-	client, err = NewClient("unknown://", "")
+	client, err = NewClient("unknown://")
 	assert.Nil(t, client)
 	assert.Error(t, err)
 	assert.Equal(t, ErrUnknownClient, err)