… memory_client_test.go