Made unicode ASCIIsation optional and disabled by default as it requires significant (20) memory allocations. Additionally made some improvements to bucket metrics generation to reduce memory allocations from 2-6 to 0.

**Connection DSN also sligtly changed, see details in README**

Master branch:

```
$ go test -bench . -benchmem
goos: darwin
goarch: amd64
pkg: github.com/hellofresh/stats-go/bucket
BenchmarkNewHTTPRequest-8                 	 1000000	      1339 ns/op	     304 B/op	      34 allocs/op
BenchmarkHttpRequest_MetricWithSuffix-8   	  500000	      3796 ns/op	    1184 B/op	      80 allocs/op
BenchmarkHttpRequest_MetricTotal-8        	  500000	      3011 ns/op	     672 B/op	      72 allocs/op
BenchmarkNewPlain-8                       	 2000000	       833 ns/op	     144 B/op	      22 allocs/op
BenchmarkPlain_Metric-8                   	10000000	       185 ns/op	      48 B/op	       3 allocs/op
BenchmarkPlain_MetricWithSuffix-8         	 3000000	       538 ns/op	     288 B/op	       6 allocs/op
BenchmarkPlain_MetricTotal-8              	10000000	       149 ns/op	      32 B/op	       2 allocs/op
BenchmarkPlain_MetricTotalWithSuffix-8    	 3000000	       501 ns/op	     256 B/op	       5 allocs/op
PASS
ok  	github.com/hellofresh/stats-go/bucket	15.257s
```

This branch

```
$ go test -bench . -benchmem
goos: darwin
goarch: amd64
pkg: github.com/hellofresh/stats-go/bucket
BenchmarkNewHTTPRequest-8                	 3000000	       489 ns/op	     175 B/op	       6 allocs/op
BenchmarkNewHTTPRequest_unicode-8        	 1000000	      1361 ns/op	     304 B/op	      34 allocs/op
BenchmarkNewPlain-8                      	10000000	       229 ns/op	      64 B/op	       2 allocs/op
BenchmarkNewPlain_unicode-8              	 2000000	       825 ns/op	     144 B/op	      22 allocs/op
BenchmarkPlain_Metric-8                  	50000000	        28.0 ns/op	       0 B/op	       0 allocs/op
BenchmarkPlain_MetricWithSuffix-8        	20000000	        63.0 ns/op	       0 B/op	       0 allocs/op
BenchmarkPlain_MetricTotal-8             	100000000	        21.1 ns/op	       0 B/op	       0 allocs/op
BenchmarkPlain_MetricTotalWithSuffix-8   	30000000	        57.4 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/hellofresh/stats-go/bucket	15.065s
```